refactor: extract session header components
